pyproject.toml
[build-system]
requires = ["flit_core >=3.7,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "sdk_entrepot_gpf"
authors = [
    {name="Valentin Sasyan", email="valentin.sasyan@ign.fr"},
]
readme="README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
]
dynamic = ["version"]
description = "SDK Python pour simplifier l'utilisation de l'API Entrepôt de la Géoplateforme."

dependencies = [
    "requests>=2.6",
    "configparser",
    "types-requests",
    "jsonschema",
    "jsonc-parser",
    "toml",
    "types-toml",
    "python-dateutil",
    "types-python-dateutil",
    "pyotp",
    "requests_toolbelt",
]

[project.optional-dependencies]
test = [
    "black<23",
    "pylint==2.17",
    "mypy==0.981",
    "requests_mock",
    "coverage",
]
doc = [
    "mkdocs-material==9.*",
    "mkdocstrings[python]",
]

[project.urls]
Source = "https://github.com/Geoplateforme/sdk_entrepot"
Documentation = "https://geoplateforme.github.io/sdk-entrepot/"

[project.scripts]
realpython = "sdk_entrepot_gpf.__main__:main"

[tool.black]
line-length = 200
force-exclude = '(\.git|\.vscode|\.toml$)'

[tool.coverage.run]
omit = ["*/tests/*", "sdk_entrepot_gpf/io/OutputManager.py", "*/__main__.py", "*/__init__.py"]
